+++ NEWS.md
@@ -1,3 +1,7 @@
+# tidypopgen 0.4.0
+* store chromosomes as factors, and remove redundant column storing
+  chromosome as integer
+
 # tidypopgen 0.3.3
 * only use the FBM backend (rather than a complete bigsnpr object)
 * fix a few minor bugs